lab 2 funcs
