+++ mitdeeplearning/__init__.py
@@ -1,2 +1,4 @@
-import mitdeeplearning.lab1
 import mitdeeplearning.util
+
+import mitdeeplearning.lab1
+import mitdeeplearning.lab2
