+++ setup.py
@@ -11,7 +11,6 @@ def get_dist(pkgname):
 install_deps = [
     'numpy',
     'regex',
-    'progressbar',
     'tqdm',
 
 ]
@@ -22,13 +21,13 @@ if get_dist('tensorflow>='+tf_ver) is None and get_dist('tensorflow_gpu>='+tf_ve
 setup(
   name = 'mitdeeplearning',         # How you named your package folder (MyLib)
   packages = ['mitdeeplearning'],   # Chose the same as "name"
-  version = '0.3.7',      # Start with a small number and increase it with every change you make
+  version = '0.3.8',      # Start with a small number and increase it with every change you make
   license='MIT',        # Chose a license from here: https://help.github.com/articles/licensing-a-repository
   description = 'Official software labs for MIT Introduction to Deep Learning (http://introtodeeplearning.com)',   # Give a short description about your library
   author = 'Alexander Amini',                   # Type in your name
   author_email = 'introtodeeplearning-staff@mit.edu',      # Type in your E-Mail
   url = 'http://introtodeeplearning.com',   # Provide either the link to your github or to your website
-  download_url = 'https://github.com/aamini/introtodeeplearning_labs/archive/v0.3.7.tar.gz',    # I explain this later on
+  download_url = 'https://github.com/aamini/introtodeeplearning_labs/archive/v0.3.8.tar.gz',    # I explain this later on
   keywords = ['deep learning', 'neural networks', 'tensorflow', 'introduction'],   # Keywords that define your package best
   install_requires=install_deps,
   classifiers=[
